p83 - text parser moved from parser.cpp to Core/ScriptParser.* - particle system significantly redesigned: most old-style particle effects done with single function and controlled with data - "cl_newfx->integer" -> NEW_FX macro - some "float time" -> "double time" to increase precision with large numbers
OpenGLDrv:
- gl_light.cpp: DiffuseLight(): added lighting for surfaces without normal info (sprites)
- removed inverseSawtoothTable[], shader parser can create ANY "inverse" wave function
- gl_shadersyntax.h: support for "map ./texture.exe" syntax
- QGL_SwapBuffers() will no more throw fatal error when single frame cannot be swapped

git-svn-id: svn://localhost@99 2093b545-fd6f-4e46-8590-e32825573395

+++ client/cl_cin.cpp
@@ -31,7 +31,7 @@ typedef struct
 	char	imageName[MAX_OSPATH];		// used for "map image.pcx"
 
 	int		frame;
-	float	frameTime;					// == cls.realtime for current cinematic frame; float precision for correct
+	double	frameTime;					// == cls.realtime for current cinematic frame; float precision for correct
 										// video/audio synchronization; why: S_RawSamples() will not restart sound, but sound
 										// buffer will be filled every 1000/CIN_SPEED ~ 71.4 msecs; when round this value to
 										// 71 or 72 will get desynchronization about 0.5/71 == 0.007 => ~0.42 sec per minute ...)
